Babel: Fix parsing of update TLVs with FLAG_ROUTER_ID set.
It's technically in spec (though not terribly useful) to send update
TLVs with an address family other than IP6 and the ROUTER_ID flag set.
Babeld produces such TLVs on startup and shutdown, making Bird spit out
errors. This accepts the TLVs for parsing (but they will still be
discarded as useless when they are actually handled later on).

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>