Babel: Garbage collect prefixes sooner when export filters change.
When an export filter is changed to remove some of the exported routes,
the route would cease to be announced immediately, but only partly
garbage collected after the short expiry interval. This makes sure that
the babel_source entry is also expired after a short hold time, making
the entire prefix be eligible for garbage collection sooner.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>