babel: Add support for dual-stack v4/v6 operation
This adds support for dual-stack v4/v6 operation to the Babel protocol. Routing
messages will be exchanged over IPv6, but IPv4 routes can be carried in the
messages being exchanged. This matches how the reference Babel
implementation (babeld) works.

The nexthop address for v4 can be configured per interface, and will default to
the first available IPv4 address on the given interface. For symmetry, a
configuration option to configure the IPv6 nexthop address is also added.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>