sysdep: Add wrapper to get random bytes
The Babel authentication code added by a subsequent commit needs a way to
get random bytes for generating nonces.

This patch adds a wrapper function in sysdep to get random bytes, and the
required checks in configure.ac to select how to do it. The configure
script tries, in order, getrandom(), getentropy() and reading from
/dev/urandom.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>