nest: Add Blake2s and Blake2b hash functions
The Babel MAC authentication draft recommends implementing Blake2s as one
of the supported algorithms. In order to achieve do this, add the blake2b
and blake2s hash functions for MAC authentication. The hashing function
implementations are the reference implementations from blake2.net, which
are imported wholesale along with the required wrapper functions to conform
to the algorithm abstraction in lib/.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>