nest: Add Blake2s and Blake2b hash functions
The Babel MAC authentication draft recommends implementing Blake2s as one
of the supported algorithms. In order to achieve do this, add the blake2b
and blake2s hash functions for MAC authentication. The hashing function
implementations are the reference implementations from blake2.net, which
are imported wholesale along with the required wrapper functions to conform
to the algorithm abstraction in lib/.

The Blake2 algorithms allow specifying an arbitrary output size, and the
Babel MAC spec says to implement Blake2s with 128-bit output. To satisfy
this, we add two different variants of each of the algorithms, one using
the default size (256 bits for Blake2s, 512 bits for Blake2b), and one
using half the default output size.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>