babel: Refactor packet parsing code for reuse in authentication checks
The subsequent commit will add authentication support to the Babel
protocol, which also requires parsing the packet TLVs. To make this easier,
this commit refactors the packet parsing code by adding a helper macro to
loop over TLVs, and generalising the babel_read_tlv function so it can be
reused for parsing authentication data.

The reusable definitions are moved into a header file which can be included
from the authentication code.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>