mac_test: Add tests for blake2s and blake2b
This imports the blake2-kat.h header with test vector output from the blake
reference implementation, and adds tests to mac_test.c to compare the
output of the Bird MAC algorithm implementations with that reference
output.

Since the reference implementation only has test vectors for the full
output size, there are no tests for the smaller-sized output variants.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>