nest: Allow specifying security keys as hex bytes as well as strings
This adds support for specifying a password in raw hexadecimal bytes form,
via the 'key' keyword. The result is the same whether a password is
specified as a quoted string or a hex-encoded byte string, this just makes
it more convenient to input high-entropy byte strings as MAC keys.

This means that the config string:

password "test" { algorithm hmac sha256; };

is equivalent to:

key 74:65:73:74 { algorithm hmac sha256; };

or

key 74657374 { algorithm hmac sha256; };

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>